Fixes #4
The initial catalogue construction in trim_cat now implements a nearby
pair check, and will remove entries in each of the source and template
catalogues which are close.

This helps with having multiple crossmatches in the matching stage.
Currently both objects in the nearby pair are removed. This could
cause large regions of the source or template to have no matches and
allow the splines to be inaccurate. An alternative is to only remove
the fainter of the close pair, thereby weighting the alignment to the
brighter portion of pairs (i.e. brighter portion of a disjointed PSF).

An additional check is in place when matching detections for other
nearby sources, although this should largely be a redundanct safety
check with the above changes.